Moved project helper to general helpers sub module
